+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImpl.java
@@ -669,7 +669,7 @@ import java.util.concurrent.CopyOnWriteArraySet;
   }
 
   @Override
-  public void setPlaybackParameters(@Nullable PlaybackParameters playbackParameters) {
+  public void setPlaybackParameters(PlaybackParameters playbackParameters) {
     if (playbackParameters == null) {
       playbackParameters = PlaybackParameters.DEFAULT;
     }